requirements.txt
opencv-python
docopt
tensorflow-gpu
keras
gym
scikit-image
statsmodels
numba
redis